Sat, Dec 3, 2022, 8:17 AM +00:00

+++ filter-more/helpers.c
@@ -36,7 +36,7 @@ void reflect(int height, int width, RGBTRIPLE image[height][width])
 // Blur image
 void blur(int height, int width, RGBTRIPLE image[height][width])
 {
-    int redsum, bluesum, greensum;
+    float redsum, bluesum, greensum;
     for (int i = 0; i < height; i++)
     {
         for (int j = 0; j < width; j++)
